requirements.txt
streamlit
numpy
matplotlib
scipy
scikit-learn
google-generativeai
dotenv